Sf branch

+++ UDP/UDPServerSocket.h
@@ -1,11 +1,13 @@
 #ifndef UDPSERVERSOCKET_H
 #define UDPSERVERSOCKET_H
 
+#include "UDPSocket.h"
+
 class UDPServerSocket : public UDPSocket {
 public:
     UDPServerSocket();
 
-    bool initializeServer(char *_myAddr, int _myPort);
+    bool initializeServer(char *_myAddr, uint16_t _myPort);
 
     ~UDPServerSocket();
 };
